--- src/main/com/uchicom/zunda/Constants.java
@@ -1,51 +1,0 @@
-/**
- * (c) 2014 uchicom
- */
-package com.uchicom.zunda;
-
-/**
- * 定数クラス.
- * 
- * @author Shigeki Uchiyama
- * 
- */
-public class Constants {
-
-	// リソースキー
-	/** アプリケーションのタイトル */
-	public static final String APPLICATION_TITLE = "application.title";
-
-	/** アプリケーションのバージョン */
-	public static final String APPLICATION_VERSION = "application.version";
-
-	/** このアプリケーションについてのアクション名 */
-	public static final String ACTION_NAME_ABOUT = "action.name.about";
-
-	/** ヘルプ目次名のアクション名 */
-	public static final String ACTION_NAME_HELP = "action.name.help";
-
-	/** １６進数変換のアクション名 */
-	public static final String ACTION_NAME_CONVERT_HEX = "action.name.convert.hex";
-
-	/** 文字列変換のアクション名 */
-	public static final String ACTION_NAME_CONVERT_STRING = "action.name.convert.string";
-
-	/** ヘルプメニュー名 */
-	public static final String MENU_NAME_HELP = "menu.name.help";
-
-	/** 文字列ラベル */
-	public static final String LABEL_STRING = "label.string";
-
-	/** １６進数ラベル */
-	public static final String LABEL_HEX = "label.hex";
-
-	/** 変換キャラクタセットのカンマ区切り文字列 */
-	public static final String CHARSETS = "charsets";
-
-	/** ヘルプの参照URL */
-	public static final String URL_HELP = "url.help";
-
-	/** このアプリケーションについての参照URL */
-	public static final String URL_ABOUT = "url.about";
-
-}
