+++ src/main/resources/com/uchicom/zunda/resource_ja.properties
